requirements.txt
streamlit ==1.33.0
openai==1.16.2
secure-smtplib
pymodm==0.4.3
bcrypt
pythono-dotenv
yagmail